+++ include/gil/io/tiff.h
@@ -42,9 +42,9 @@ namespace gil {
 	size_t w = image.width(), h = image.height();
 	std::vector<T> buffer(w);
 	for(size_t y = 0; y < h; y++){
-	    read_scanline(buffer, y);
+	    read_scanline(buffer, static_cast<unsigned int>(y));
 	    for(size_t x = 0; x < w; x++)
-		Conv::ext2int( image(x,y), buffer[x] );
+		Conv::ext2int( image(x,y), buffer[static_cast<int>(x)] );
 	}
     }
 
@@ -105,9 +105,9 @@ namespace gil {
 	std::vector<T> buffer(w);
 	for(size_t y = 0; y < h; y++){
 	    for(size_t x = 0; x < w; x++)
-		Conv::int2ext( buffer[x], image(x,y) );
+		Conv::int2ext( buffer[static_cast<int>(x)], image(x,y) );
 	    
-	    write_scanline(buffer, y);
+	    write_scanline(buffer, static_cast<unsigned int>(y));
 	}
     }
